extract spec runner to allow to have a second set
